# src/back-end/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/item": {
      "get": {
        "operationId": "ItemController_findWithPattern",
        "parameters": [
          {
            "name": "pattern",
            "required": false,
            "in": "query",
            "description": "Search pattern for item name",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "unit",
            "required": false,
            "in": "query",
            "description": "Unit of the item",
            "schema": {
              "example": "Cái",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The items have been successfully fetched.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The pattern is invalid."
          }
        },
        "tags": [
          "Item"
        ]
      }
    },
    "/item/{id}": {
      "put": {
        "operationId": "ItemController_updateItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The item has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemUpdateDto"
                }
              }
            }
          }
        },
        "tags": [
          "Item"
        ]
      }
    },
...
  },
  "info": {
    "title": "Agency me example",
    "description": "The Agency-me API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "[just text field] Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "ItemDto": {
        "type": "object",
        "properties": {
          "mathang_id": {
            "type": "string",
            "description": "The ID of the item",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "ten": {
            "type": "string",
            "description": "The name of the item",
            "example": "Item 1"
          },
          "don_gia": {
            "type": "number",
            "description": "The price of the item",
            "example": 100
          },
          "don_vi_tinh": {
            "type": "string",
            "description": "The quantity of the item",
            "example": "Cái"
          },
          "ky_hieu": {
            "type": "string",
            "description": "The symbol or code of the item",
            "example": "$"
          }
        },
        "required": [
          "mathang_id",
          "ten",
          "don_gia",
          "don_vi_tinh"
        ]
      },
      "ItemUpdateDto": {
        "type": "object",
        "properties": {
          "don_gia": {
            "type": "number",
            "description": "The price of the item",
            "example": 100
          }
        },
        "required": [
          "don_gia"
        ]
      },
...
    }
  }
}